refactor(wash-runtime): build wasm test fixtures via xtask
Move wasm test-fixture compilation out of crates/wash-runtime/build.rs
into a new root-level xtask crate, invoked explicitly as
`cargo xtask build-fixtures`.

Building fixtures from build.rs meant `cargo test`/`cargo build` spawned
nested `cargo build` invocations ("cargo building cargo"), with the
jobserver/lock contention and rebuild storms that come with it. Pulling
the work into an explicitly-invoked binary keeps fixture compilation off
the normal build path entirely, so build.rs is now pure-Rust proto
codegen.

- Add /xtask (excluded from default-members) with a build-fixtures task;
  one batched `cargo build` per WASI target rather than one spawn per
  fixture, failing loud on any error. Drops an ambient CARGO_TARGET_DIR
  so staging finds the artifacts on the bench hosts.
- Reduce build.rs to proto codegen (306 -> 78 lines); remove the unused
  WORKSPACE_ROOT env.
- Run `cargo xtask build-fixtures` before tests in wash.yml and
  codeql.yml, and in scripts/bench/run-bench.sh (guarded so
  compare-bench.sh still works against pre-xtask refs).
- Pin the WASI reactor adapter to 45, in lockstep with wasmtime.
- Document the flow in CONTRIBUTING, the wash-runtime README, and a new
  tests/fixtures/README.

Based on the approach and prototype in #5085 by @dman-os (one of several
variations alongside #4998 and #5089). Thanks for working through the
design and the explicit-fixture-building idea.
