+++ .github/workflows/wash.yml
@@ -161,6 +161,11 @@ jobs:
           repo-token: ${{ secrets.GITHUB_TOKEN }}
       - name: Setup Rust
         uses: ./.github/actions/setup-rust
+      # Compile the wash-runtime wasm test fixtures. The test/bench targets
+      # `include_bytes!` these, so they must exist before `cargo test`/`cargo
+      # bench` below. Targets wasm32-wasip2/wasip1 come from setup-rust.
+      - name: Build test fixtures
+        run: cargo xtask build-fixtures
       - name: Build
         run: cargo build
       - name: Test
