+++ xtask/Cargo.toml
@@ -1,0 +1,20 @@
+# Repo task runner (the cargo-xtask pattern). Not published; invoked via
+# the `cargo xtask` alias in /.cargo/config.toml. Kept out of
+# `default-members` so normal `cargo build`/`cargo test` never pull it in.
+[package]
+name = "xtask"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+publish = false
+
+[dependencies]
+anyhow = { workspace = true, default-features = true }
+# Workspace clap is default-features=false; xtask needs std on top.
+clap = { workspace = true, features = ["std"] }
+wit-component = { workspace = true }
+# Pins the WASI reactor adapter in lockstep with our wasmtime version
+# (used to componentize P3 core modules). Keep the major in sync with
+# the `wasmtime`/`wit-component` versions in the workspace manifest.
+wasi-preview1-component-adapter-provider = "45"
